+++ src/layers/XRLayer/shader-modules/channel-intensity.glsl
@@ -90,6 +90,10 @@ vec3 process_channel_intensity(float intensity, vec3 colorValues, int channelInd
   return hsv_to_rgb(hsvCombo);
 }
 
+vec4 apply_transparency(vec3 color, bool useTransparentColor, vec4 transparentColor, float opacity){
+  return vec4(color, (color == transparentColor.rgb && useTransparentColor) ? 0.0 : opacity);
+}
+
 vec4 colormap(float intensity, float opacity) {
   return vec4(COLORMAP_FUNCTION(min(1.0,intensity)).xyz, opacity);
 }